requirements.txt
streamlit==1.28.1
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
plotly==5.17.0
tensorflow==2.13.0
opencv-python==4.8.0.76
pillow==10.0.0
matplotlib==3.7.2
seaborn==0.12.2
